Sapien/requirements.txt
sapien==2.2.2
numpy==1.26.4
imageio==2.37.0
imageio-ffmpeg==0.6.0
opencv-python-headless==4.10.0.84
Pillow==10.4.0
PyYAML==6.0.2
tqdm==4.67.1
